fix some warnings and minor things like naming and XCode configs (#175)

+++ Overlay-Graphics/Overlay-Graphics/TBAudioLevelMeter.m
@@ -35,7 +35,7 @@
     self.frame = frame;
 }
 
-- (void)setLevel:(float)level
+- (void)setLevel:(double)level
 {
     _level = level;
     [self setNeedsDisplay];
